Refactor PDF splitter into modular classes
